It works now!
Improved control so that calculations are done using the apparent wind
direction.
Big work on control of jib and mainsail to make sure they catch as much
wind as possible, and only get blown to the correct direction.
Corrected torque calculations from the sails and rudder by correctly
calculating the centers of force application on each.
Improved drawing of jib, mainsail, and rudder to be drawn with the
correct points fixed in place.