Physics constants improvement, auto-tune, better sail control
Physics constants were improved so that the boat is not ridiculously
efficient at catching the wind.
Auto-tuning for PID was written, but performed very poorly and was
commented out for future experimentation.
Better PI sail control added with better use of points of sail and
coordination with rudder control. Rudder is PD unless the sails max out
their integral term. Then the rudder is PID. This should result in the
most efficient turning and course control.